+++ Android/Sunshine/S03.01-Exercise-RecyclerView/app/src/main/java/com/example/android/sunshine/ForecastAdapter.java
@@ -1,0 +1,48 @@
+package com.example.android.sunshine;
+
+import android.support.v7.widget.RecyclerView;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.TextView;
+
+public class ForecastAdapter extends RecyclerView.Adapter<ForecastAdapter.ForecastAdapterViewHolder> {
+
+    private String[] mWeatherData;
+
+    public ForecastAdapter() {
+
+    }
+
+    public void setWeatherData(String[] weatherData) {
+        mWeatherData = weatherData;
+        notifyDataSetChanged();
+    }
+
+    @Override
+    public ForecastAdapterViewHolder onCreateViewHolder(ViewGroup viewGroup, int viewType) {
+        View view = LayoutInflater.from(viewGroup.getContext())
+                .inflate(R.layout.forecast_list_item, viewGroup, false);
+
+        return new ForecastAdapterViewHolder(view);
+    }
+
+    @Override
+    public void onBindViewHolder(ForecastAdapterViewHolder holder, int position) {
+        holder.mWeatherTextView.setText(mWeatherData[position]);
+    }
+
+    @Override
+    public int getItemCount() {
+        return (mWeatherData == null) ? 0 : mWeatherData.length;
+    }
+
+    class ForecastAdapterViewHolder extends RecyclerView.ViewHolder {
+        public final TextView mWeatherTextView;
+
+        public ForecastAdapterViewHolder(View view) {
+            super(view);
+            mWeatherTextView = (TextView) view.findViewById(R.id.tv_weather_data);
+        }
+    }
+}